refactor: move busboy logic to util image file
