+++ backend/src/controllers/user.ts
@@ -1,4 +1,3 @@
-import busboy from "busboy";
 import { NextFunction, Request, Response } from "express";
 import { validationResult } from "express-validator";
 import admin from "firebase-admin";
@@ -10,7 +9,7 @@ import { ValidationError } from "../errors/validation";
 import { Image } from "../models/image";
 import UserModel from "../models/user";
 import { firebaseAdminAuth } from "../util/firebase";
-import { saveImage } from "../util/image";
+import { handleImageParsing } from "../util/image";
 import validationErrorParser from "../util/validationErrorParser";
 
 // Define the type for req.body
@@ -125,7 +124,7 @@ export type SaveImageRequest = {
   };
 };
 
-type CustomRequest = Request & {
+export type CustomRequest = Request & {
   userId: string;
   rawBody?: Buffer;
 };
@@ -133,72 +132,8 @@ type CustomRequest = Request & {
 export const editPhoto = (req: Request, res: Response, nxt: NextFunction) => {
   try {
     const customReq = req as CustomRequest;
-    let previousImageId = "";
-    //req.userId is assigned in verifyAuthToken middleware
-    const userId: string = customReq.userId;
 
-    const bb = busboy({ headers: req.headers });
-
-    let fileBuffer = Buffer.alloc(0);
-    bb.on("field", (fieldname, val) => {
-      if (fieldname === "previousImageId") {
-        previousImageId = val;
-      }
-    });
-    bb.on("file", (name, file, info) => {
-      const { filename, mimeType } = info;
-
-      file
-        .on("data", (data) => {
-          fileBuffer = Buffer.concat([fileBuffer, data]);
-        })
-        .on("close", () => {
-          const saveImageRequest: SaveImageRequest = {
-            body: {
-              previousImageId,
-              userId,
-            },
-            file: {
-              buffer: fileBuffer,
-              originalname: filename,
-              mimetype: mimeType,
-              size: fileBuffer.length,
-            },
-          };
-
-          // Validate file in form data
-          try {
-            const acceptableTypes = ["image/jpeg", "image/png", "image/webp"];
-            if (!acceptableTypes.includes(mimeType)) {
-              throw ValidationError.IMAGE_UNSUPPORTED_TYPE;
-            }
-
-            // Check file size (2MB limit)
-            const maxSize = 2 * 1024 * 1024;
-            if (fileBuffer.length > maxSize) {
-              throw ValidationError.IMAGE_EXCEED_SIZE;
-            }
-
-            saveImage(saveImageRequest)
-              .then((savedImageId) => {
-                res.status(200).json(savedImageId);
-              })
-              .catch((error) => {
-                console.error("Error saving image:", error);
-                nxt(error); // Properly forward the error
-              });
-          } catch (error) {
-            console.error("Error parsing form with Busboy:", error);
-            nxt(error); // Properly forward the error
-          }
-        });
-    });
-
-    if (customReq.rawBody) {
-      bb.end(customReq.rawBody);
-    } else {
-      customReq.pipe(bb);
-    }
+    handleImageParsing(customReq, res, nxt);
   } catch (e) {
     console.log(e);
     nxt(e);
